moved Auth to 'hooks' folder
