+++ src/index.js
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import ReactDOM from "react-dom";
 
-import { ProvideAuth } from "./components/Auth";
+import { ProvideAuth } from "./hooks/Auth";
 import App from "./App";
 
 const rootElement = document.getElementById("root");
